fake_profile_detection_system/requirements.txt
numpy
pandas
scikit-learn
xgboost
lightgbm
tensorflow
keras
opencv-python
matplotlib
seaborn
plotly
streamlit
transformers
torch
networkx
PyYAML
nltk
textblob
deepface
